requirements.txt
flask==0.10.1
flask-script
wtforms==1.0.5
flask-wtf==0.9.3
Flask-Mail
https://github.com/mrjoes/flask-admin/tarball/master
flask-sqlalchemy
flask-whooshalchemy
gevent
